+++ proto/cosmos/vesting/v1beta1/tx.proto
@@ -6,6 +6,15 @@ import "cosmos/base/v1beta1/coin.proto";
 
 option go_package = "github.com/cosmos/cosmos-sdk/x/auth/vesting/types";
 
+
+// Msg defines the bank Msg service.
+service Msg {
+  // CreateVestingAccount defines a method that enables creating a vesting
+  // account.
+  rpc CreateVestingAccount(MsgCreateVestingAccount) returns (MsgCreateVestingAccountResponse);
+}
+
+
 // MsgCreateVestingAccount defines a message that enables creating a vesting
 // account.
 message MsgCreateVestingAccount {
@@ -19,3 +28,6 @@ message MsgCreateVestingAccount {
   int64 end_time = 4 [(gogoproto.moretags) = "yaml:\"end_time\""];
   bool  delayed  = 5;
 }
+
+// MsgCreateVestingAccountResponse defines the Msg/CreateVestingAccount response type.
+message MsgCreateVestingAccountResponse { }
